Added support for different repository types

+++ scripts/builder.py
@@ -1,6 +1,5 @@
 #!/usr/bin/python3
 
-import argparse
 import json
 import os
 import shutil
@@ -8,8 +7,13 @@ import sys
 
 from component import Component
 from config import CLADE_CC, CLADE_INTERCEPT, COMPONENT_BUILDER, TAG_CLADE_CONF, \
-    TAG_MAKE_COMMAND, TAG_FAIL_IF_FAILURE, CLADE_BASE_FILE, CLADE_DEFAULT_CONFIG_FILE, CLADE_WORK_DIR, \
-    DEFAULT_INSTALL_DIR
+    TAG_MAKE_COMMAND, TAG_FAIL_IF_FAILURE, CLADE_BASE_FILE, CLADE_DEFAULT_CONFIG_FILE, CLADE_WORK_DIR
+
+REPOSITORY_GIT = "git"
+REPOSITORY_SVN = "svn"
+REPOSITORY_NONE = None
+
+REPOSITORY_TYPES = [REPOSITORY_GIT, REPOSITORY_SVN, REPOSITORY_NONE]
 
 TAG_ENVIRON_VARS = "environment variables"
 TAG_CLEAN_SOURCES = "clean sources"
@@ -23,10 +27,15 @@ CLADE_BASIC_CONFIG = {
 
 
 class Builder(Component):
-    def __init__(self, install_dir, config, source_dir, builder_config={}):
+    def __init__(self, install_dir, config, source_dir, builder_config={}, repository=REPOSITORY_NONE):
         super(Builder, self).__init__(COMPONENT_BUILDER, config)
         self.install_dir = install_dir
         self.source_dir = source_dir
+
+        if repository not in REPOSITORY_TYPES:
+            sys.exit("Repository type '{}' is not known. Supported types of repositories are: {}".
+                     format(repository, REPOSITORY_TYPES))
+        self.repository = repository
 
         if not builder_config:
             self.is_build = False
@@ -48,30 +57,54 @@ class Builder(Component):
         self.env = self.component_config.get(TAG_ENVIRON_VARS, {})
 
     def clean(self):
-        self.logger.debug("Clear the source directory")
-        os.chdir(self.source_dir)
-        if self.clean_sources:
-            self.command_caller("rm -rf *")
-        self.command_caller("git reset --hard")
-        os.chdir(self.work_dir)
+        if self.repository:
+            self.logger.debug("Cleaning the source directory {}".format(self.source_dir))
+            os.chdir(self.source_dir)
+            if self.clean_sources:
+                self.command_caller("rm -rf *")
+            if self.repository == REPOSITORY_GIT:
+                self.command_caller("git reset --hard")
+            elif self.repository == REPOSITORY_SVN:
+                self.command_caller("svn revert --recursive .")
+            else:
+                sys.exit("Operation 'clean' is not implemented for repository type '{}'".format(self.repository))
+            os.chdir(self.work_dir)
 
     def change_branch(self, branch: str):
-        self.logger.debug("Using branch '{}' in the source directory".format(branch))
-        os.chdir(self.source_dir)
-        if self.command_caller("git checkout {}".format(branch)):
-            sys.exit("Cannot checkout to branch '{}' in the source directory".format(branch))
-        self.command_caller("git reset --hard")
-        os.chdir(self.work_dir)
+        if self.repository:
+            self.logger.debug("Using branch '{}' in the source directory {}".format(branch, self.source_dir))
+            os.chdir(self.source_dir)
+            if self.repository == REPOSITORY_GIT:
+                if self.command_caller("git checkout {}".format(branch)):
+                    sys.exit("Cannot checkout to GIT branch '{}' in the source directory".format(branch))
+                self.command_caller("git reset --hard")
+            elif self.repository == REPOSITORY_SVN:
+                if self.command_caller("svn switch ^/{}".format(branch)):
+                    sys.exit("Cannot switch to SVN branch '{}' in the source directory".format(branch))
+                self.command_caller("svn revert --recursive .")
+            else:
+                sys.exit("Operation 'change branch' is not implemented for repository type '{}'".
+                         format(self.repository))
+            os.chdir(self.work_dir)
 
     def patch(self, patch: str):
         os.chdir(self.source_dir)
         if not os.path.exists(patch):
             sys.exit("Specified patch '{}' does not exists".format(patch))
-        self.logger.debug("Applying patch '{}' to the source directory".format(patch))
-        command = "git apply --ignore-space-change --ignore-whitespace {}".format(patch)
-        if self.command_caller(command):
-            self.logger.error("Command '{}' failed".format(command))
-            sys.exit("Cannot apply patch '{}' to the source directory {}".format(patch, self.source_dir))
+        self.logger.debug("Applying patch '{}' to the source directory {}".format(patch, self.source_dir))
+        if self.repository == REPOSITORY_GIT:
+            command = "git apply --ignore-space-change --ignore-whitespace {}".format(patch)
+            if self.command_caller(command):
+                self.logger.error("Command '{}' failed".format(command))
+                sys.exit("Cannot apply patch '{}' to the source directory {}".format(patch, self.source_dir))
+        elif self.repository == REPOSITORY_SVN:
+            if self.command_caller("svn patch {}".format(patch)):
+                sys.exit("Cannot apply patch '{}' to the source directory {}".format(patch, self.source_dir))
+        elif not self.repository:
+            if self.command_caller("patch -lf -p1 < {}".format(patch)):
+                sys.exit("Cannot apply patch '{}' to the source directory {}".format(patch, self.source_dir))
+        else:
+            sys.exit("Operation 'patch' is not implemented for repository type '{}'".format(self.repository))
         os.chdir(self.work_dir)
 
     def build(self, build_commands_file: str):
@@ -109,7 +142,8 @@ class Builder(Component):
         self.logger.info("Sources has been successfully built")
 
 
-#TODO: not supported.
+# TODO: not supported anymore.
+'''
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
     parser.add_argument("--install", "-i", help="install directory", default=os.path.abspath(DEFAULT_INSTALL_DIR))
@@ -125,3 +159,4 @@ if __name__ == '__main__':
     builder.clean()
     builder.patch(options.patch)
     builder.build()
+'''
